+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.1] - 2021-10-25
 ### Added
 - mock: add `record_ddl` handler to test DDL queries.
 - mock: add `watch` and `watch_only_events` handlers to test WATCH queries.
@@ -134,7 +135,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Client::insert()` for inserting into tables.
 - `Client::query()` for selecting from tables and DDL statements.
 
-[unreleased]: https://github.com/loyd/clickhouse.rs/compare/v0.9.0...HEAD
+[unreleased]: https://github.com/loyd/clickhouse.rs/compare/v0.9.1...HEAD
+[0.9.1]: https://github.com/loyd/clickhouse.rs/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/loyd/clickhouse.rs/compare/v0.8.1...v0.9.0
 [0.8.1]: https://github.com/loyd/clickhouse.rs/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/loyd/clickhouse.rs/compare/v0.7.2...v0.8.0